feat: using GlobalConfig to access wrapping_key, only needs to be generated once for all test cases now
