+++ tomb/src/cli.rs
@@ -231,10 +231,11 @@ mod test {
         // Initialization worked
         init(&origin).await?.assert().success();
         // Assert the bucket exists now
-        let bucket = GlobalConfig::from_disk()?.get_bucket(origin);
+        let global = GlobalConfig::from_disk()?;
+        // Assert that there is always a wrapping key
+        assert!(global.wrapping_key_from_disk().is_ok());
+        let bucket = global.get_bucket(origin);
         assert!(bucket.is_some());
-        // Assert that there is still no key, because we've not packed
-        assert!(bucket.unwrap().private_key_from_disk().is_err());
         // Teardown test
         test_teardown(test_name).await
     }
