+++ tomb/src/pipelines/remove.rs
@@ -10,9 +10,11 @@ use super::error::PipelineError;
 pub async fn pipeline(origin: &Path, wnfs_path: &Path) -> Result<(), PipelineError> {
     // Global config
     let mut global = GlobalConfig::from_disk()?;
+    let wrapping_key = global.wrapping_key_from_disk()?;
     // Bucket config
     if let Some(config) = global.get_bucket(origin) {
-        let (metadata_forest, content_forest, dir, key_manager) = &mut config.get_all().await?;
+        let (metadata_forest, content_forest, dir, key_manager) =
+            &mut config.get_all(&wrapping_key).await?;
         // Attempt to remove the node
         dir.rm(
             &path_to_segments(wnfs_path)?,
@@ -24,7 +26,13 @@ pub async fn pipeline(origin: &Path, wnfs_path: &Path) -> Result<(), PipelineErr
 
         // Stores the modified directory back to disk
         config
-            .set_all(metadata_forest, content_forest, dir, key_manager)
+            .set_all(
+                &wrapping_key,
+                metadata_forest,
+                content_forest,
+                dir,
+                key_manager,
+            )
             .await?;
 
         // Update global
