+++ tomb/src/pipelines/unpack.rs
@@ -25,10 +25,12 @@ pub async fn pipeline(origin: &Path, unpacked: &Path) -> Result<(), PipelineErro
     info!("🚀 Starting unpacking pipeline...");
 
     let mut global = GlobalConfig::from_disk()?;
+    let wrapping_key = global.wrapping_key_from_disk()?;
 
     if let Some(config) = global.get_bucket(origin) {
         // Load metadata
-        let (metadata_forest, content_forest, dir, key_manager) = &mut config.get_all().await?;
+        let (metadata_forest, content_forest, dir, key_manager) =
+            &mut config.get_all(&wrapping_key).await?;
         let metadata = &config.metadata;
         let content = &config.content;
 
@@ -101,14 +103,18 @@ pub async fn pipeline(origin: &Path, unpacked: &Path) -> Result<(), PipelineErro
             content,
         )
         .await?;
-
         // Set all
         config
-            .set_all(metadata_forest, content_forest, dir, key_manager)
+            .set_all(
+                &wrapping_key,
+                metadata_forest,
+                content_forest,
+                dir,
+                key_manager,
+            )
             .await?;
         global.update_config(&config)?;
         global.to_disk()?;
-
         Ok(())
     } else {
         Err(PipelineError::Uninitialized)
