+++ tomb/src/utils/mod.rs
@@ -2,6 +2,8 @@
 pub mod custom_fclones_logger;
 /// This module contains code designed to traverse directory structure and get a packing plan for files, including deduplication
 pub mod grouper;
+/// Utils specific to the Pack pipeline
+pub mod pack;
 /// This module contains code designed to traverse directory structure and get a packing plan for directories and symlinks.
 pub mod spider;
 /// This module contains testing fns
